Added support for GME (Game_Music_Emu).
This makes it possible for LÖVE to play music files for NES, SNES, MegaDrive,
and many more. It's pretty cool.

The feature is disabled until each platform is ready to link with GME. Packages
exist for Debian and Arch, and it compiled without any issues in MSVC++, so
I don't expect this to a problem for anyone.